zecpath-project/Backend/requirements.txt
# Basic requirements for environment setup
requests==2.31.0
django==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
python-magic==0.4.27
Pillow==10.4.0

# Resume parsing
pdfplumber==0.10.3
python-docx==1.1.0
PyPDF2==3.0.1

# Performance & Load Testing
locust==2.15.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# Async Task Processing
celery==5.3.4
redis==5.0.1

# AI/Voice APIs (Day 35)
openai==1.3.0
twilio==8.10.0